+++ source/ProcessManager.Orchestrations.Tests/Integration/Processes/BRS_021/ElectricalHeatingCalculation/V1/MonitorOrchestrationUsingClientsScenario.cs
@@ -15,44 +15,46 @@
 using Energinet.DataHub.Core.TestCommon;
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model;
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model.OrchestrationInstance;
+using Energinet.DataHub.ProcessManager.Client;
 using Energinet.DataHub.ProcessManager.Client.Extensions.DependencyInjection;
 using Energinet.DataHub.ProcessManager.Client.Extensions.Options;
-using Energinet.DataHub.ProcessManager.Client.Tests.Fixtures;
 using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_021.ElectricalHeatingCalculation.V1.Model;
+using Energinet.DataHub.ProcessManager.Orchestrations.Tests.Fixtures;
+using Energinet.DataHub.ProcessManager.Orchestrations.Tests.Fixtures.Extensions;
 using FluentAssertions;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Xunit.Abstractions;
 
-namespace Energinet.DataHub.ProcessManager.Client.Tests.Integration.BRS_021.ElectricalHeatingCalculation.V1;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.Tests.Integration.Processes.BRS_021.ElectricalHeatingCalculation.V1;
 
 /// <summary>
 /// Test case where we verify the Process Manager clients can be used to start a
 /// calculation orchestration (with no input parameter) and monitor its status during its lifetime.
 /// </summary>
-[Collection(nameof(ObsoleteProcessManagerClientCollection))]
-public class MonitorCalculationUsingClientsScenario : IAsyncLifetime
+[Collection(nameof(OrchestrationsAppCollection))]
+public class MonitorOrchestrationUsingClientsScenario : IAsyncLifetime
 {
-    public MonitorCalculationUsingClientsScenario(
-        ObsoleteProcessManagerClientFixture fixture,
+    public MonitorOrchestrationUsingClientsScenario(
+        OrchestrationsAppFixture fixture,
         ITestOutputHelper testOutputHelper)
     {
         Fixture = fixture;
         Fixture.SetTestOutputHelper(testOutputHelper);
 
         var services = new ServiceCollection();
-        services.AddScoped(_ => CreateInMemoryConfigurations(new Dictionary<string, string?>()
+        services.AddInMemoryConfiguration(new Dictionary<string, string?>
         {
             [$"{ProcessManagerHttpClientsOptions.SectionName}:{nameof(ProcessManagerHttpClientsOptions.GeneralApiBaseAddress)}"]
                 = Fixture.ProcessManagerAppManager.AppHostManager.HttpClient.BaseAddress!.ToString(),
             [$"{ProcessManagerHttpClientsOptions.SectionName}:{nameof(ProcessManagerHttpClientsOptions.OrchestrationsApiBaseAddress)}"]
                 = Fixture.OrchestrationsAppManager.AppHostManager.HttpClient.BaseAddress!.ToString(),
-        }));
+        });
         services.AddProcessManagerHttpClients();
         ServiceProvider = services.BuildServiceProvider();
     }
 
-    private ObsoleteProcessManagerClientFixture Fixture { get; }
+    private OrchestrationsAppFixture Fixture { get; }
 
     private ServiceProvider ServiceProvider { get; }
 
@@ -122,12 +124,5 @@ public class MonitorCalculationUsingClientsScenario : IAsyncLifetime
                 CancellationToken.None);
 
         orchestrationInstancesGeneralSearch.Should().Contain(x => x.Id == orchestrationInstanceId);
-    }
-
-    private IConfiguration CreateInMemoryConfigurations(Dictionary<string, string?> configurations)
-    {
-        return new ConfigurationBuilder()
-            .AddInMemoryCollection(configurations)
-            .Build();
     }
 }
